to lorentzvector

+++ processing/interface/feat_comp.hh
@@ -10,8 +10,9 @@
 // ROOT
 #include <Math/VectorUtil.h>
 #include <Math/LorentzVector.h>
+#include <Math/PxPyPzM4D.h>
 
-using LorentzVector = ROOT::Math::LorentzVector<ROOT::Math::PxPyPzM4D<double> >;
+using LorentzVector = ROOT::Math::LorentzVector<ROOT::Math::PxPyPzM4D<float>>;
 
 class FeatComp {
 	/* Class for computing requested features for final-state LorentzVectors*/
